refactor atests
